// generated/src/app/api-models/vendor-address.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class VendorAddress {
  seqVendAddress: number;
  seqVendId: number;
  name2: string;
  addressLine1: string;
  addressLine2: string;
  city: string;
  state: string;
  zipCode: string;
  county: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  userDefined1: string;
  userDefined2: string;
  userDefinedDate: Date;
  seqVendAddrBillOvrd: number;
  country: string;
  primaryAddress: string;
  userDate2: Date;
}